@@ -1,10 +1,11 @@
 name: flutter_easy
 description: A common Flutter package.
-version: 3.9.7
+version: 3.10.0
 homepage: https://github.com/OctMon/flutter_easy
 
 environment:
-  sdk: ">=3.2.3 <4.0.0"
+  sdk: '>=3.3.0 <4.0.0'
+  flutter: ">=3.27.0"
 
 dependencies:
   flutter:
@@ -20,7 +21,7 @@ dependencies:
   # ===============================================================================
 
   dart_art: ^0.1.0+1
-  image: ^4.4.0
+  image: ^4.5.2
   intl: ^0.19.0
 
   # ===============================================================================
@@ -57,7 +58,7 @@ dependencies:
   # desc：包装NSUserDefaults（在iOS上）和SharedPreferences（在Android上），为简单数据提供持久存储。
   # 数据异步保存到磁盘。 这两个平台都不能保证在返回后将写操作持久化到磁盘，并且该插件不得用于存储关键数据。
   # ===============================================================================
-  shared_preferences: ^2.3.3
+  shared_preferences: ^2.3.5
   shared_preferences_platform_interface: ^2.4.1
   shared_preferences_web: ^2.4.2
 
@@ -101,7 +102,7 @@ dependencies:
   # pub: https://pub.dev/packages/extended_image
   # git: https://github.com/fluttercandies/extended_image
   # ———————————————————————————————————————————————————————————————————————————————
-  extended_image: ^9.0.7
+  extended_image: ^9.1.0
   # ===============================================================================
 
   # ===============================================================================
