@@ -6,11 +6,12 @@ Color? baseDefaultShimmerBaseColor;
 Color? baseDefaultShimmerHighlightColor;
 
 class BaseShimmer extends StatelessWidget {
-  const BaseShimmer({super.key,
-    required this.visible,
-    required this.child,
-    this.baseColor,
-    this.highlightColor});
+  const BaseShimmer(
+      {super.key,
+      required this.visible,
+      required this.child,
+      this.baseColor,
+      this.highlightColor});
 
   final bool visible;
   final Widget child;
@@ -23,10 +24,10 @@ class BaseShimmer extends StatelessWidget {
       return Shimmer.fromColors(
         baseColor: baseColor ??
             baseDefaultShimmerBaseColor ??
-            appTheme(context).primaryColor.withOpacity(0.3),
+            appTheme(context).primaryColor.withValues(alpha: 0.3),
         highlightColor: highlightColor ??
             baseDefaultShimmerHighlightColor ??
-            appTheme(context).primaryColor.withOpacity(0.6),
+            appTheme(context).primaryColor.withValues(alpha: 0.6),
         child: IgnorePointer(child: child),
       );
     } else {
