@@ -76,7 +76,7 @@ Future<Result> postAPI(
 ///
 /// 发送请求并解析远程服务器返回的result对应的实体类型
 ///
-/// <T>: 要解析的实体类名(需要自动转换时必须要加)
+/// T: 要解析的实体类名(需要自动转换时必须要加)
 /// baseUrl: 主机地址
 /// path: 请求路径
 /// data: 请求参数
